refactor: split metamodel folder
