+++ Cargo.toml
@@ -1,16 +1,3 @@
-[package]
-name = "concerto_core"
-version = "0.1.0"
-edition = "2021"
-authors = ["Accord Project"]
-description = "Rust implementation of Concerto Core for structural and semantic metamodel validation"
-
-[dependencies]
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-thiserror = "1.0"
-regex = "1.10"
-semver = "1.0"
-log = "0.4"
-lazy_static = "1.4"
-chrono = "0.4"
+[workspace]
+resolver = "3"
+members = ["concerto-metamodel", "concerto-core"]
