+++ concerto-core/src/lib.rs
@@ -1,15 +1,14 @@
 // Main library file for concerto_core
 
 // Define modules
+pub mod error;
+pub mod introspect;
+pub mod metamodel_validation;
 pub mod model_file;
 pub mod model_manager;
-pub mod error;
-pub mod validation;
-pub mod introspect;
-pub mod util;
-pub mod metamodel;
-pub mod metamodel_validation;
 pub mod traits;
+pub mod util;
+pub mod validation;
 
 // Re-export the main components
 pub use model_file::ModelFile;
@@ -17,13 +16,12 @@ pub use model_manager::ModelManager;
 pub use traits::*;
 
 // Metamodel types exports
-pub use metamodel::concerto_metamodel_1_0_0::{
-    Model, Declaration, Import,
-    Property, Decorator, TypeIdentifier
+pub use concerto_metamodel::concerto_metamodel_1_0_0::{
+    Declaration, Decorator, Import, Model, Property, TypeIdentifier,
 };
 
-pub use validation::Validate;
 pub use error::ConcertoError;
+pub use validation::Validate;
 
 /// Version of the Concerto core library
 pub const VERSION: &str = env!("CARGO_PKG_VERSION");
