+++ concerto-metamodel/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "concerto-metamodel"
+version = "3.12.8"
+edition = "2024"
+
+[dependencies]
+serde = { version = "1.0", features = ["derive"] }
+chrono = "0.4"
