+++ concerto-metamodel/src/lib.rs
@@ -1,0 +1,7 @@
+mod metamodel;
+
+pub use metamodel::concerto;
+pub use metamodel::concerto_1_0_0;
+pub use metamodel::concerto_decorator_1_0_0;
+pub use metamodel::concerto_metamodel_1_0_0;
+pub use metamodel::utils;
